Expanded echo command
The echo command can now be used by other people. I wrote a list of 
three additional users I know who now have access to the echo command. 
Only these users right now, as a pilot program, basically so I can test 
to see if I should expand it further to all administrators within 
various servers. As part of this change, I also set it to report to me 
when one of these people actually uses the command. This is mainly so I 
know if someone is abusing it.