Added mod role code for new channels
When a new channel is created, it should contain the correct permissions 
for the mute and timeout roles, so the function works correctly. This 
does happen by default, as long as the new channel is created within a 
category that already has the permissions properly set. However, if the 
category perms are not properly set; or the channel is created outside 
the categories, as would happen if the new channel actually is a 
category, this doesn't happen. So, I wrote some code to do this 
automatically by adding the correct permissions if it's necessary.