Added a message for server bump
So everyone knows Midnight is properly recording server bumps, I decided 
to set a message saying the bumps are being accurately recorded, and to 
report the member's current bump score. For the purposes of good 
grammar, I also added a function to get the ordinal number of a given 
integer, because if someone bumped the server 11 times, their next bump 
is the '12th bump' not the '12 bump'. The latter makes it sound like 
there are 12 bumps being recorded at that moment, when there is only 
one.