Added bump leaderboard
To help incentivize server bumping, I added the ability to post a list 
of all users who bumped the server, ranked by most bumps to least bumps. 
Basically, it's a server bump leaderboard. Most of the code is 
duplicated from the rule posting code, but obviously altered so it works 
with this new idea.