Added role control feature
The new role control feature is designed to ensure a member only picks 
one role out of a given group. This might be because a bunch of roles 
are mutually exclusive, like stating gender or species, or higher roles 
make the lower ones redundant, like in the case of colour roles. It 
prevents ridiculous bloat if using a self-assigned-roles system, and 
some users decide to go nuts with it.