Moved command checks
In order to speed up command processing, I moved the check that 
determines which command has been called to the on_message event 
handler. That way it doesn't have to peek into every single function, 
running the same parsing code 50 times, just to check if it's the right 
function. It was ridiculously inefficient. This change does create a few 
redundancies I have to clear up within the individual functions, since 
it still checks the command within the function, meaning the command is 
checked twice. But the net result of this first step should be some 
significantly more efficient code. Which is ideal.