Added special code for Ultimate Pony Verse
A friend asked me to automate some role assignments for him, based on a 
user's display name. So I wrote the code specifically for his server. I 
wanted it to be a bit more customizable, and usable by every server, but 
I haven't figured out a way to implement it yet. So for now, it's 
custom-written specifically for his specific server. I decided to place 
the code in a special module that I'll reserve for custom and 
specifically-built code. It may contain more at a later date, but for 
now, this is enough.