Cleaned up some code in Midnight.py
Continuing my work from a few days ago 
(490753290b447985c34a715772c069c4f5ce6c26), I decided to start going 
through all the command functions, and removing the redundant command 
checks, now that they're being run in the event handler. As part of 
this, I also set it so the initial parsing code that isolates the 
command from any arguments that may follow is done in the parent handler 
as well, and the result of that passed into the function, rather than 
run the code twice, once in the event handler, and again in the 
function. As well, there was an unnecessary line of code in the rolecall 
function that I removed that stripped whitespace, which is already done 
to the commandArgs before they are passed into the function. And as a 
final kick, since I was cleaning stuff up anyway, I also removed the 
IS_ECHO_ENABLED constant from the code, since there's really no reason 
to disable it at this point. The echo command is not going anywhere.