Rewrote emoji censor code
Rewrote the code for the emoji censor to be more accurate and concise. I 
changed the code that searches for Discord emojis to work with regular 
expressions. And I placed that regular expression in the regular 
expression that already checks for standard emojis. This simplifies the 
whole code block and makes it more efficient. I also replaced the 
standard emoji regex with a new one I found through the Discord.py 
discord server. This one should be more accurate.