Moved client object back to Midnight.py
In a reversion of an earlier commit 
(46c59146158056a83d6ad6f75f1a839b80ed1686), which was made around the 
time I split the project into multiple files, I moved the client 
initialization from the Utilities module back to the startup file. This 
is mainly so I can keep some logical consistency. It just makes sense to 
have the client object initialized in the main startup file, rather than 
the Utilities module. I fixed the problem of having certain functions 
access the client object by passing it into each of the functions as an 
argument.