Edited Postgres instructions
The instructions to install Postgres did not specify a version number, 
as I realized when I had to reinstall everything on my development 
machine after replacing the hard drive. So I investigated, learned which 
version is running on my production machine, and decided to specify that 
9.6 was the Postgres version I was developing on, and therefore the 
minimum version of Postgres that should be used. I also changed the 
specified database name to be all lowercase, because I had some 
compatibility issues in the past regarding case sensitivity. It's 
simpler if the database name is just all lowercase.