Dealt with invites with no inviters
To deal with a weird glitch I don't quite understand, sometimes invites 
appear with no inviter on Discord. I don't know why, or how to prevent 
it. So instead I rewrote the bot to just deal with it. Any invite with 
no known inviter, the inviter is recorded as a user with an id of '0'. 
This'll handle most of the problems that are likely to pop up because of 
this. Might even help me track down the cause since the invites are 
still being recorded in the database.