Altered import in Constants file
I decided to alter the 'import os' command to only import the one 
function that's actually used in the Constants file. This is partly 
because there's no need to import the whole library when you're only 
using a single part of it, but it's also because I don't want the 'os' 
library to be imported in every file, which is what is currently 
happening. This fixes it so I have to explicitly import the library into 
the files that use it. Such as the Economy file, which didn't.