Added ambient audio feature
Added a new feature that does nothing but play ambient audio in one of 
the voice channels. I set it up to loop the audio, and disconnect when 
everyone leaves the voice channel. Simple enough. Audio tracks must be 
in Opus format and are retrieved from a specified directory called 
'Ambience'. I'm not including the tracks in the repository, because I'm 
not sure about the copyright, and the files can be massive.