Built special verification system
Created a verification system that works as a coded call and response 
message. All one has to do is complete a poem from the season 2 finale 
of My Little Pony. Coded to work only on one specific server. Should 
prevent non-pony fans from raiding said server. Originally it was built 
for one friend, but when that friend basically threw the work back in my 
face, I retooled it so it would work on another friend's server. Might 
end up retooling it to be more universal, but for now, this works.