Modified voice connection handling
I had a few problems with the voice connection cache not being accurate, 
to say the least. So I modified the code to check the bot's voice state 
in a given server, rather than just check if the connection cache has 
anything, and assume an empty cache means no voice connection. In case 
there is a voice connection, but the connection cache is empty, it loops 
through the voice clients stored by the API to find the relevant 
connection. I'm still using the cache, because the loop, while more 
reliable, is significantly inefficient. This should fix all the problems 
I've been having with the voice connections.