Altered how import statements are handled
Since I split the project into multiple files (f9c5c94141b12bbbd562dfdd99ba2b263e6090d5) I've been bugged by the import statements I used to link the subfiles to the main file. Rather than import the files, I imported everything within the files. At the time I did this because it made moving the code to the new files simpler, and I didn't need to rewrite too much. But because of this, in all calls to external constants and functions, the files they originate in are not in any way specified, or even that they're external in the first place. This also means that function and variable names have to be unique across all files, which can be problematic. So I altered the import statements to import the entire files, so the files have to be specified when I call the relevant functions. This is primarily a stylistic change, but I think it improves matters, mainly for the sake of clarity. I'm also planning to move some specific code around that works better if it's module-specific. This will simplify that process.