Fixed dependency loop
There was a problem with the most recent change that added the prefix command (f2bf88bf6ceca1cfa9a63d2f985ad9603561ac67), that was preventing database connections from properly starting. Basically, the DATABASE_URL constant wasn't being properly initialized before being accessed for some reason. By my best guess, this was happening because the Utilities file tried to import the Constants file, which also tried to import the Utilities file, and that created a sort of feedback loop of some kind where neither knew which one to load first. I fixed it by just changing how the Constants file is imported in the Utilities file, from importing the individual items to importing the whole file all at once. I'm guessing the items didn't exist when the file was imported, and that is what caused the problem.