Updated bot to use Intents
There was a change to how bots interact with Discord, that I was only recently made aware of. Most actions are now restricted unless I specifically enable them in both my bot's code and the Discord Developer Portal. I updated the code, the relevant libraries, and the installation instructions. This should solve a lot of the problems I've been having.