Redid handling of role deletions
Added some code to deal with the deletion of any role relied on by 
certain features. When a role is deleted, I wrote a single function that 
goes through all the necessary database tables and searches for any use 
of that role and removes it. This is to prevent certain exceptions from 
occurring in the event a role is deleted without the necessary feature 
being disabled first.