Updated error handling
To improve error handling, I added a specific message to be sent to server users for errors involving inadequate permissions. It's more detailed than the generic cheekiness it previously sent for any and all errors, (and is still sent for all other errors). This now gives them the minimum amount of information they need to fix the problem. I still need to add more detailed error messages for this, but I'd need to program those in every necessary function individually. This is a simple way to fulfill the basic function, which is enough for now.