Reformatted copyright boilerplates
Just did a minor formatting change to the copyright boilerplate, mainly because I started looking at some things about python formatting, and thought the indent should be after the comment symbol, and not before. This is also due to the fact that I found a new Python IDE, which flagged the indentation as a problem. This new IDE will likely mean I will be making several stylistic changes to the project in the near future, which should be interesting.